detector/requirements.txt
flask==3.0.3
requests==2.31.0
pyyaml==6.0.1
psutil==5.9.8
